yourlabs/project_template/requirements.txt
# those require compiling toolchain, you can remove them if your OS
# is broken ...
pil
pylibmc

django-debug-toolbar
django-debug-toolbar-template-timings
django-cache-panel

# classic
-e git+https://github.com/pdr/django-responsive-admin.git#egg=responsive_admin
django-extensions
ipython
ipdb
werkzeug
django-haystack
whoosh
django-reversion
django-modeltranslation
south
django-autocomplete-light

# for docs
sphinx
pycco

# for production deployment
django-compressor
jsmin
cssmin
BeatifulSoup
